backend/app/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pillow==10.1.0
numpy==1.24.3
mediapipe==0.10.9
scikit-learn==1.3.2
joblib==1.3.2
opencv-python-headless==4.8.1.78
tensorflow==2.15.0  # Comenta si da problemas
tensorflow-cpu==2.15.0  # Alternativa más ligera
python-dotenv==1.0.0
aiofiles==23.2.1